(feat): update invoice breakdown to support varied project performance periods and refactor some files to the right package (#281)
* (feat): update invoice breakdown to support varied project performance periods and refactor some files to the right package

* fix formatting
