+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/module/BusinessProjectModule.java
@@ -36,14 +36,14 @@ import com.axelor.apps.businessproject.service.approvalitem.ApprovalItemManageme
 import com.axelor.apps.businessproject.service.approvalitem.ApprovalItemManagementServiceImpl;
 import com.axelor.apps.businessproject.service.config.BusinessProjectConfigService;
 import com.axelor.apps.businessproject.service.config.BusinessProjectConfigServiceImpl;
-import com.axelor.apps.businessproject.service.extracharges.InvoiceBreakdownDisplayService;
-import com.axelor.apps.businessproject.service.extracharges.InvoiceBreakdownDisplayServiceImpl;
-import com.axelor.apps.businessproject.service.extracharges.InvoiceBreakdownPrintService;
-import com.axelor.apps.businessproject.service.extracharges.InvoiceBreakdownPrintServiceImpl;
 import com.axelor.apps.businessproject.service.extraexpense.ExtraExpenseInvoiceService;
 import com.axelor.apps.businessproject.service.extraexpense.ExtraExpenseInvoiceServiceImpl;
 import com.axelor.apps.businessproject.service.extraexpense.ExtraExpenseLineService;
 import com.axelor.apps.businessproject.service.extraexpense.ExtraExpenseLineServiceImpl;
+import com.axelor.apps.businessproject.service.invoice.InvoiceBreakdownDisplayService;
+import com.axelor.apps.businessproject.service.invoice.InvoiceBreakdownDisplayServiceImpl;
+import com.axelor.apps.businessproject.service.invoice.InvoiceBreakdownPrintService;
+import com.axelor.apps.businessproject.service.invoice.InvoiceBreakdownPrintServiceImpl;
 import com.axelor.apps.businessproject.service.invoice.InvoiceMergingServiceBusinessProjectImpl;
 import com.axelor.apps.businessproject.service.invoice.InvoicePrintBusinessProjectService;
 import com.axelor.apps.businessproject.service.invoice.InvoicePrintBusinessProjectServiceImpl;
