+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/web/InvoicingProjectController.java
@@ -28,8 +28,8 @@ import com.axelor.apps.businessproject.db.repo.InvoicingProjectRepository;
 import com.axelor.apps.businessproject.exception.BusinessProjectExceptionMessage;
 import com.axelor.apps.businessproject.service.InvoicingProjectService;
 import com.axelor.apps.businessproject.service.ProjectGenerateInvoiceService;
-import com.axelor.apps.businessproject.service.extracharges.InvoiceBreakdownDisplayService;
-import com.axelor.apps.businessproject.service.extracharges.InvoiceBreakdownPrintService;
+import com.axelor.apps.businessproject.service.invoice.InvoiceBreakdownDisplayService;
+import com.axelor.apps.businessproject.service.invoice.InvoiceBreakdownPrintService;
 import com.axelor.apps.project.db.Project;
 import com.axelor.apps.project.db.repo.ProjectRepository;
 import com.axelor.i18n.I18n;
